DEV: Integrate bcolz and SQLite readers/writers for FFC.

+++ zipline/data/equities.py
@@ -1,5 +1,5 @@
 from numpy import (
-    float32,
+    float64,
     uint32,
 )
 
@@ -11,8 +11,8 @@ from zipline.data.dataset import (
 
 class USEquityPricing(DataSet):
 
-    open = Column(float32)
-    high = Column(float32)
-    low = Column(float32)
-    close = Column(float32)
+    open = Column(float64)
+    high = Column(float64)
+    low = Column(float64)
+    close = Column(float64)
     volume = Column(uint32)
